+++ packages/entity-graph-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @prometheus-ags/entity-graph-core
 
+## 3.0.4
+
 ## 3.0.3
 
 ### Patch Changes